+++ src/routes/vessels.js
@@ -1,14 +1,20 @@
 const datasets = require('../data/datasets');
 const vessels = require('../data/vessels');
 const tracks = require('../data/tracks');
+const log = require('../data/log');
 
 const loadDataset = async (req, res, next) => {
   try {
     const datasetId = req.swagger.params.dataset.value;
+
+    log.debug(`Looking up dataset ${datasetId}`);
     const dataset = await datasets.get(datasetId);
     if (!dataset) {
+      log.debug('Dataset not found');
       return res.sendStatus(404);
     }
+
+    log.debug('Dataset found', dataset);
     req.dataset = dataset;
     return next();
   } catch (err) {
@@ -25,10 +31,13 @@ module.exports = (app) => {
         offset: req.swagger.params.offset.value,
       };
 
-      const results = await vessels(req.dataset).search(query);
+      log.debug('Querying vessels search index');
+      const results = await vessels(req.dataset)
+        .search(query);
 
+      log.debug(`Returning ${results.entries.length} / ${results.total} results`);
       return res.json(results);
-    } catch(error) {
+    } catch (error) {
       return next(error);
     }
   });
@@ -37,10 +46,13 @@ module.exports = (app) => {
     try {
       const vesselId = req.swagger.params.vesselId.value;
 
-      const result = await vessels(req.dataset).get(vesselId);
+      log.debug(`Looking up vessel information for vessel ${vesselId}`);
+      const result = await vessels(req.dataset)
+        .get(vesselId);
 
+      log.debug('Returning vessel information');
       return res.json(result);
-    } catch(error) {
+    } catch (error) {
       return next(error);
     }
   });
@@ -50,10 +62,13 @@ module.exports = (app) => {
     try {
       const vesselId = req.swagger.params.vesselId.value;
 
-      const result = await tracks(req.dataset).forVessel(vesselId);
+      log.debug(`Looking up track for vessel ${vesselId}`);
+      const result = await tracks(req.dataset)
+        .forVessel(vesselId, req.swagger.params.features.value);
 
+      log.debug(`Returning track for vessel ${vesselId}`);
       return res.json(result);
-    } catch(error) {
+    } catch (error) {
       return next(error);
     }
   });