+++ src/services/elasticsearch.js
@@ -1,6 +1,6 @@
-const config = require('../config');
 const { Client } = require('elasticsearch');
+const config = require('../config');
 
 module.exports = new Client({
   host: config.elasticsearch.server,
-})
+});